Test framework for gp parallel retrieve interface (phrase 3)
In order to skipping token id difference when comparing expected output and real
output (token id in SQL and in the result of query), we should set some
matchsubs section like initfile (--gpd_init param for gpdiff.pl).

Sample in test case:
 TOKEN1=` echo "${RAW_STR}" | awk \'NR==3\' | awk \'{print $1}\'` && export MATCHSUBS="${MATCHSUBS}${NL}m/${TOKEN1}/${NL}s/${TOKEN1}/token_id1/${NL}" && echo "${RAW_STR}" ': SELECT token,hostname,status FROM GP_ENDPOINTS WHERE cursorname='c1'

The env var ${MATCHSUBS} is used to store the matchsubs section so that we can
store it into initfile when this test case file is finished
executing.